pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "satellite-anomaly-detection"
version = "1.0.0"
description = "LSTM Autoencoder + 1D CNN pipeline for satellite telemetry anomaly detection and fault classification"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Sachchit Vekaria" }]
keywords = ["anomaly-detection", "satellite", "lstm", "autoencoder", "cnn", "time-series"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.21",
    "pandas>=1.3",
    "scipy>=1.7",
    "torch>=2.0",
    "scikit-learn>=1.0",
    "matplotlib>=3.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4"]
notebooks = ["jupyter>=1.0", "plotly>=5.0"]

[project.scripts]
sat-anomaly = "sat_anomaly.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
